+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:miocardio_paciente/login.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:miocardio_paciente/pages/login.dart';
+import 'package:miocardio_paciente/generated/i18n.dart';
 
 void main() => runApp(MyApp());
 
@@ -8,6 +10,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      localizationsDelegates: [
+          S.delegate,
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate
+        ],
+      supportedLocales: S.delegate.supportedLocales,
       title: 'Miocardio Pediatra',
       theme: ThemeData(
         primarySwatch: Colors.blue,
