+++ lib/pages/home.dart
@@ -1,21 +1,24 @@
 import 'package:flutter/material.dart';
+import 'package:miocardio_paciente/generated/i18n.dart';
 
-class Info extends StatefulWidget {
-  InfoState createState() => InfoState();
+
+class Home extends StatefulWidget{
+  HomeState createState() => HomeState();
+
 }
+class HomeState extends State<Home>{
 
-class InfoState extends State<Info> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Color.fromRGBO(253, 224, 224, 1),
-      body: ListView.builder(
+            backgroundColor: Color.fromRGBO(253, 224, 224, 1),
+            body: ListView.builder(
           itemCount: 1,
           padding: EdgeInsets.all(5),
           itemBuilder: (BuildContext context, int index) => Container(
                 padding: EdgeInsets.only(top: 30, bottom: 10),
                 child: Text(
-                  'Info',
+                  S.of(context).pagetitleHome,
                   textAlign: TextAlign.center,
                   style: TextStyle(
                     color: Colors.black,
@@ -24,6 +27,6 @@ class InfoState extends State<Info> {
                   ),
                 ),
               )),
-    );
+      );
   }
 }
