+++ lib/pages/info.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:miocardio_paciente/generated/i18n.dart';
 
-class ReminderPage extends StatefulWidget {
-  ReminderPageState createState() => ReminderPageState();
+class Info extends StatefulWidget {
+  InfoState createState() => InfoState();
 }
 
-class ReminderPageState extends State<ReminderPage> {
+class InfoState extends State<Info> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -15,7 +16,7 @@ class ReminderPageState extends State<ReminderPage> {
           itemBuilder: (BuildContext context, int index) => Container(
                 padding: EdgeInsets.only(top: 30, bottom: 10),
                 child: Text(
-                  'Lembretes',
+                  S.of(context).pagetitleInfo,
                   textAlign: TextAlign.center,
                   style: TextStyle(
                     color: Colors.black,
